cart controller added

+++ StyleHubApi/StyleHubApi/models/product.cs
@@ -8,12 +8,12 @@ namespace StyleHubApi.Models
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public string? Size { get; set; }
         public int Review { get; set; }
         public bool In_stock { get; set; }
         public string? Badge { get; set; }
-        public double? OldPrice { get; set; }
+        public decimal? OldPrice { get; set; }
         public double Rating { get; set; }
         public string? Color { get; set; }
         public List<string> Images { get; set; } = new();
